Apply tuikit

+++ src/previewer.rs
@@ -1,6 +1,6 @@
-use event::{Event, EventSender};
-use ansi::AnsiString;
-use libc;
+use crate::ansi::AnsiString;
+use crate::event::{Event, EventSender};
+use nix::libc;
 
 use std::env;
 use std::io::Read;
@@ -12,8 +12,8 @@ use std::thread;
 
 pub struct PreviewInput {
     pub cmd: String,
-    pub lines: u16,
-    pub columns: u16,
+    pub lines: usize,
+    pub columns: usize,
 }
 
 struct PreviewThread {
